Restructure file structure for components
